Fixing the bug to show new account requests  when admin is logged in
