+++ src/main/webapp/security/protected/admin/activate_yubikey.xhtml
@@ -33,7 +33,7 @@
 
             <p:column>
               <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"> 
-                #{peopleAdministration.selectedYubikyUser.title} #{peopleAdministration.selectedYubikyUser.fname} #{peopleAdministration.selectedYubikyUser.lname} 
+                 #{yubikeyActivator.selectedYubikyUser.fname} #{yubikeyActivator.selectedYubikyUser.lname} 
               </h3>
             </p:column>
 
@@ -43,7 +43,7 @@
 
             <p:column>
               <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"> 
-                #{peopleAdministration.selectedYubikyUser.activated}
+                #{yubikeyActivator.selectedYubikyUser.activated}
               </h3>
             </p:column>
 
@@ -52,7 +52,7 @@
             </p:column>
 
             <p:column>
-              <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"> #{peopleAdministration.selectedYubikyUser.email}  </h3>
+              <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"> #{yubikeyActivator.selectedYubikyUser.email}  </h3>
             </p:column>
 
             <p:column>
@@ -60,7 +60,7 @@
             </p:column>
 
             <p:column>
-              <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"> #{peopleAdministration.selectedYubikyUser.mobile}  </h3>
+              <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"> #{yubikeyActivator.selectedYubikyUser.mobile}  </h3>
             </p:column>
 
             <p:column>
@@ -68,7 +68,7 @@
             </p:column>
 
             <p:column>
-              <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"> #{peopleAdministration.selectedYubikyUser.organization.orgName}  </h3>
+              <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"> #{yubikeyActivator.selectedYubikyUser.organization.orgName}  </h3>
             </p:column>
 
             <p:column>
@@ -76,7 +76,7 @@
             </p:column>
 
             <p:column>
-              <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"> #{peopleAdministration.selectedYubikyUser.orcid}  </h3>
+              <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"> #{yubikeyActivator.selectedYubikyUser.orcid}  </h3>
             </p:column>
 
 
@@ -96,7 +96,7 @@
 
             <p:column>
               <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;">              
-              #{peopleAdministration.selectedYubikyUser.fname} #{peopleAdministration.selectedYubikyUser.lname} 
+              #{yubikeyActivator.selectedYubikyUser.fname} #{yubikeyActivator.selectedYubikyUser.lname} 
 
               </h3>
             </p:column>
@@ -106,7 +106,7 @@
 
             <p:column>
               <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"  >
-                #{peopleAdministration.selectedYubikyUser.address.address2} #{peopleAdministration.selectedYubikyUser.address.address3} 
+                #{yubikeyActivator.selectedYubikyUser.address.address2} #{yubikeyActivator.selectedYubikyUser.address.address3} 
               </h3>
             </p:column>
 
@@ -117,7 +117,7 @@
 
             <p:column>
               <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"> 
-                #{peopleAdministration.selectedYubikyUser.address.postalcode} 
+                #{yubikeyActivator.selectedYubikyUser.address.postalcode} 
               </h3>
             </p:column>
 
@@ -128,7 +128,7 @@
 
             <p:column>
               <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"> 
-               #{peopleAdministration.selectedYubikyUser.address.city}
+               #{yubikeyActivator.selectedYubikyUser.address.city}
               </h3>
             </p:column>
 
@@ -138,7 +138,7 @@
 
             <p:column>
               <h3 style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"> 
-                #{peopleAdministration.selectedYubikyUser.address.country} 
+                #{yubikeyActivator.selectedYubikyUser.address.country} 
               </h3>
             </p:column>
 
@@ -154,7 +154,7 @@
                            value="Yubikey PublicId"/>
 
             <p:inputText id="pid" 
-                         value="#{peopleAdministration.pubid}" 
+                         value="#{yubikeyActivator.pubid}" 
                          required="true"
                          requiredMessage="Enter the Yubikey  Public Id" 
                          size="40">
@@ -168,7 +168,7 @@
                            style="color: rgba(0, 0, 0, 0.54); font-weight: 400;"
                            value="Yubikey Secret"/>
             <p:inputText id="secret"
-                         value="#{peopleAdministration.secret}" 
+                         value="#{yubikeyActivator.secret}" 
                          required="true"
                          requiredMessage="Enter the Yubikey Secret Code" 
                          size="40">
@@ -182,7 +182,7 @@
           </h:panelGrid>
 
           <div align="center">
-            <p:commandButton  type="submit" value="Register" action="#{peopleAdministration.activateYubikey()}"/>
+            <p:commandButton  type="submit" value="Register" action="#{yubikeyActivator.activateYubikey()}"/>
             <p:spacer width="10px"></p:spacer>
             <p:commandButton  type="submit" value="Cancel" action="userMgmt" immediate="true" />
           </div>
