+++ src/main/webapp/security/protected/admin/admin_profile.xhtml
@@ -283,7 +283,7 @@
 
           <p:column>
             <h3  style="color: rgba(0, 0, 0, 0.54); font-weight: 400;">
-              <p:outputLabel value="Last PasswordChange"/>
+              <p:outputLabel value="Last Password Change"/>
             </h3> 
           </p:column>
